pyproject.toml
[project]
name = "rag-native"
version = "0.1.0"
description = "RAG system for researchers to search and query their document collections"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    # Document Processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.6.0",
    "llama-parse>=0.5.0",
    # Embeddings & Vector Store
    "openai>=1.12.0",
    "chromadb>=0.4.22",
    # Retrieval
    "rank-bm25>=0.2.2",
    # Frontend
    "streamlit>=1.31.0",
    "requests>=2.31.0",
    # Utilities
    "python-multipart>=0.0.9",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "cohere>=5.0.0",
    "pymupdf>=1.26.7",
    "gunicorn>=23.0.0",
    # Cloud Storage (Optional)
    "supabase>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config", "ui"]
